Refactor: Update package structure and API call
- Moved `TransactionResponse.kt` to the `com.fintrack.app.data.response` package.
- Updated imports in `TransactionApiService.kt`, `NavigationActivity.kt`, and `LainnyaFragment.kt` to reflect the package change.
- Modified `BudgetRepository.kt` to include a Bearer token in the `getAllBudgets` API call.
- Corrected the package declaration in `LainnyaFragment.kt`.

+++ app/src/main/java/com/fintrack/app/data/request/BudgetPayload.kt
@@ -2,5 +2,5 @@ package com.fintrack.app.data.request
 
 data class BudgetPayload(
     val category: String,      // Nama kategori
-    val amountLimit: Int       // Jumlah maksimal pengeluaran
+    val amountLimit: Double,       // Jumlah maksimal pengeluaran
 )
